Move derivative destroy into its own package
