+++ services-camel/src/main/java/edu/unc/lib/dl/services/camel/destroy/DestroyObjectsRouter.java
@@ -32,26 +32,6 @@ public class DestroyObjectsRouter extends RouteBuilder {
     @BeanInject(value = "destroyObjectsProcessor")
     private DestroyObjectsProcessor destroyObjectsProcessor;
 
-    @BeanInject(value = "binaryInfoProcessor")
-    private BinaryInfoProcessor binaryInfoProcessor;
-
-    @BeanInject(value = "destroySmallThumbnailProcessor")
-    private DestroyDerivativesProcessor destroySmallThumbnailProcessor;
-
-    @BeanInject(value = "destroyLargeThumbnailProcessor")
-    private DestroyDerivativesProcessor destroyLargeThumbnailProcessor;
-
-    @BeanInject(value = "destroyAccessCopyProcessor")
-    private DestroyDerivativesProcessor destroyAccessCopyProcessor;
-
-    @BeanInject(value = "destroyFulltextProcessor")
-    private DestroyDerivativesProcessor destroyFulltextProcessor;
-
-    private static final String IMAGE_MIMETYPE_PATTERN = "^(image.*$|application.*?(photoshop|psd)$)";
-    private static final String TEXT_MIMETYPE_PATTERN = "^(text/|application/pdf|application/msword"
-            + "|application/vnd\\.|application/rtf|application/powerpoint"
-            + "|application/postscript).*$";
-
     @Override
     public void configure() throws Exception {
         onException(Exception.class)
@@ -64,24 +44,5 @@ public class DestroyObjectsRouter extends RouteBuilder {
             .routeId("CdrDestroyObjects")
             .log(DEBUG, "Received destroy objects message")
             .bean(destroyObjectsProcessor);
-
-        from("{{cdr.destroy.derivatives.stream.camel}}")
-            .routeId("CdrDestroyDerivatives")
-            .log(DEBUG, "Received destroy derivatives message")
-            .process(binaryInfoProcessor)
-            .choice()
-                .when(simple("${headers[CdrMimeType]} regex '" + IMAGE_MIMETYPE_PATTERN + "'"))
-                    .to("direct:image.derivatives.destroy")
-                .when(simple("${headers[CdrMimeType]} regex '" + TEXT_MIMETYPE_PATTERN + "'"))
-                    .to("direct:fulltext.derivatives.destroy")
-            .end();
-
-        from("direct:fulltext.derivatives.destroy")
-            .bean(destroyFulltextProcessor);
-
-        from("direct:image.derivatives.destroy")
-            .bean(destroySmallThumbnailProcessor)
-            .bean(destroyLargeThumbnailProcessor)
-            .bean(destroyAccessCopyProcessor);
     }
 }
