+++ services-camel/src/main/java/edu/unc/lib/dl/services/camel/destroyDerivatives/BinaryInfoProcessor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.unc.lib.dl.services.camel.destroy;
+package edu.unc.lib.dl.services.camel.destroyDerivatives;
 
 import static edu.unc.lib.dl.services.camel.util.CdrFcrepoHeaders.CdrBinaryMimeType;
 import static edu.unc.lib.dl.services.camel.util.CdrFcrepoHeaders.CdrBinaryPidId;
